requirements.txt
spotipy==2.25.1
python-dotenv==1.1.1
playwright==1.55.0
google-generativeai==0.8.5
pydantic==2.12.2
Flask==3.1.2
requests==2.32.5
apscheduler==3.11.0